refact: move account to core.account
