+++ core/sample/src/main/java/com/jeanbarrossilva/mastodonte/core/sample/profile/Profile.extensions.kt
@@ -1,7 +1,7 @@
 package com.jeanbarrossilva.mastodonte.core.sample.profile
 
 import com.jeanbarrossilva.mastodonte.core.profile.Profile
-import com.jeanbarrossilva.mastodonte.core.sample.toot.sample
+import com.jeanbarrossilva.mastodonte.core.sample.account.sample
 import com.jeanbarrossilva.mastodonte.core.toot.Author
 
 /** [Profile] returned by [sample]'s getter. **/
