+++ core/sample/src/main/java/com/jeanbarrossilva/mastodonte/core/sample/profile/edit/SampleEditableProfile.kt
@@ -1,9 +1,9 @@
 package com.jeanbarrossilva.mastodonte.core.sample.profile.edit
 
+import com.jeanbarrossilva.mastodonte.core.account.Account
 import com.jeanbarrossilva.mastodonte.core.profile.edit.EditableProfile
 import com.jeanbarrossilva.mastodonte.core.profile.edit.Editor
 import com.jeanbarrossilva.mastodonte.core.sample.profile.SampleProfile
-import com.jeanbarrossilva.mastodonte.core.toot.Account
 import java.net.URL
 
 /**
