+++ feature/profile-details/src/main/java/com/jeanbarrossilva/mastodonte/feature/profiledetails/ProfileDetails.kt
@@ -42,6 +42,7 @@ import com.jeanbarrossilva.loadable.Loadable
 import com.jeanbarrossilva.loadable.list.ListLoadable
 import com.jeanbarrossilva.loadable.list.toSerializableList
 import com.jeanbarrossilva.loadable.placeholder.MediumTextualPlaceholder
+import com.jeanbarrossilva.mastodonte.core.account.Account
 import com.jeanbarrossilva.mastodonte.core.profile.Profile
 import com.jeanbarrossilva.mastodonte.core.profile.edit.EditableProfile
 import com.jeanbarrossilva.mastodonte.core.profile.follow.FollowableProfile
@@ -49,7 +50,6 @@ import com.jeanbarrossilva.mastodonte.core.sample.profile.edit.sample
 import com.jeanbarrossilva.mastodonte.core.sample.profile.follow.sample
 import com.jeanbarrossilva.mastodonte.core.sample.profile.sample
 import com.jeanbarrossilva.mastodonte.core.sample.toot.samples
-import com.jeanbarrossilva.mastodonte.core.toot.Account
 import com.jeanbarrossilva.mastodonte.core.toot.Toot
 import com.jeanbarrossilva.mastodonte.feature.profiledetails.conversion.converter.followable.toStatus
 import com.jeanbarrossilva.mastodonte.feature.profiledetails.navigation.BackwardsNavigationState
